moves dff to modules
