+++ src/ophys_etl/modules/dff/schemas.py
@@ -1,15 +1,7 @@
-# Compute deltaF/F
-import time
-
-from argschema import ArgSchema, ArgSchemaParser, fields
-import multiprocessing
-import h5py
+from argschema import ArgSchema, fields
 from marshmallow import post_load
 import numpy as np
-from functools import partial
-from pathlib import Path
 
-from ophys_etl.transforms.trace_transforms import compute_dff_trace
 from ophys_etl.schemas.fields import H5InputFile
 
 
@@ -117,72 +109,3 @@ class DffJobSchema(ArgSchema):
         item["long_filter_frames"] = (
             long_frames if long_frames % 2 else long_frames + 1)
         return item
-
-
-def job_call(index: int, input_file: Path, key: str,
-             long_filter: int, short_filter: int):
-    with h5py.File(input_file, "r") as f:
-        trace = f[key][index]
-    dff, sigma_dff, small_baseline = compute_dff_trace(
-            trace, long_filter, short_filter)
-    return dff, sigma_dff, small_baseline
-
-
-class DffJob(ArgSchemaParser):
-    """
-    This is the job runner for the dF/F computation from F (fluorescence)
-    traces. The primary data input is the h5 file produced by neuropil
-    subtraction, which contains the neuropil-corrected fluorescence trace.
-    (by default in the "CF" key).
-
-    NOTE: There has historically been no data saved in the output json
-    for this job in the AllenSDK, so the output json is just a
-    placeholder (empty dictionary).
-    """
-    default_schema = DffJobSchema
-    default_output_schema = DffJobOutputSchema
-
-    def run(self):
-        # Set up file and data pointers
-        with h5py.File(self.args["input_file"], "r") as f:
-            traces_shape = f[self.args["input_dataset"]].shape
-            roi_dataset = f[self.args['roi_field']][()]
-        roi_shape = roi_dataset.shape
-
-        # Check for id mapping mismatches
-        if roi_shape[0] != traces_shape[0]:
-            raise ValueError(
-                f"Can't associate ROIs of shape {roi_dataset.shape} "
-                f"to traces of shape {traces_shape}")
-
-        job_partial = partial(job_call,
-                              input_file=Path(self.args['input_file']),
-                              key=self.args["input_dataset"],
-                              long_filter=self.args["long_filter_frames"],
-                              short_filter=self.args["short_filter_frames"])
-        with multiprocessing.Pool(self.args['n_parallel_workers']) as pool:
-            results = [i for i in pool.imap(
-                       job_partial, np.arange(traces_shape[0]), chunksize=10)]
-
-        dff, sigma_dff, small_baseline = list(zip(*results))
-
-        with h5py.File(self.args["output_file"], "w") as output_h5:
-            output_h5.create_dataset(self.args["roi_field"],
-                                     data=roi_dataset),
-            output_h5.create_dataset(self.args["output_dataset"], data=dff)
-            output_h5.create_dataset(self.args["sigma_dataset"],
-                                     data=sigma_dff)
-            output_h5.create_dataset(self.args["baseline_frames_dataset"],
-                                     data=small_baseline)
-
-        self.logger.info("Dff traces complete.")
-
-        self.output({
-            "output_file": self.args["output_file"],
-            "created_at": int(time.time())
-            }, indent=2)
-
-
-if __name__ == "__main__":    # pragma: nocover
-    dff_job = DffJob()
-    dff_job.run()
