+++ tests/transforms/test_trace_transforms.py
@@ -59,50 +59,3 @@ def test_extract_traces(frames, rois, normalize_by_roi_size, expected):
     obtained = tx.extract_traces(frames, rois, normalize_by_roi_size)
 
     assert np.allclose(obtained, expected)
-
-
-@pytest.mark.parametrize(
-    "x, expected",
-    [
-        (np.zeros(10), 0.0),    # Zeros
-        (np.ones(20), 0.0),     # All same, not zero
-        (np.array([-1, -1, -1]), 0.0),   # Negatives
-        (np.array([]), np.NaN),   # Empty
-        (np.array([0, 0, np.NaN, 0.0]), np.NaN),     # Has NaN
-        (np.array([1]), 0.0),    # Unit
-        (np.array([-1, 2, 3]), 1.4826)    # Typical
-    ]
-)
-def test_robust_std(x, expected):
-    np.testing.assert_equal(expected, tx.robust_std(x))
-
-
-@pytest.mark.parametrize(
-    "x, expected",
-    [
-        (np.array([0, 1, 2, 3, np.NaN]), np.NaN),
-        (np.array([4, 10, 4, 5, 4]), 2.5)
-    ])
-def test_noise_std(x, expected, monkeypatch):
-    monkeypatch.setattr(tx, "robust_std", lambda x: x.max())
-    monkeypatch.setattr(tx, "medfilt", lambda x, y: x/2)
-    np.testing.assert_equal(expected, tx.noise_std(x))
-
-
-def test_dff_trace(monkeypatch):
-    """
-    Notes:
-    If we don't constrain this it's very unwieldy. Not using
-    parametrization because these values need to be
-    monkeypatched thoughtfully to make a unit test work out
-    Isn't a great candidate for mock because most of the
-    logic pertains to filtering numpy arrays anyway.
-    """
-    monkeypatch.setattr(tx, "noise_std", lambda x, y: 1.0)
-    monkeypatch.setattr(tx, "medfilt", lambda x, y: x-1.0)
-    f_trace = np.array([1.1, 2., 3., 3., 3., 11.])    # 2 "small baseline"
-
-    dff, sigma, small_baseline = tx.compute_dff_trace(f_trace, 1, 1)
-    assert 2 == small_baseline
-    assert 1.0 == sigma     # monkeypatched noise_std
-    np.testing.assert_array_equal(np.ones(6), dff)
